rewrote filename handling to properly handle being given a "blah.map" filename when doing vis/light/etc which actually need a .bsp filename consolidated all the filename generation to one place in qbsp.c so that this is much easier to maintain
git-svn-id: svn://svn.icculus.org/twilight/trunk@5459 d7cf8633-e32d-0410-b094-e92efae38249

+++ hmap2/brush.c
@@ -551,7 +551,7 @@ brush_t *LoadBrush (mbrush_t *mb, int brushnum, int hullnum)
 
 	if (!brush_faces)
 	{
-		printf ("WARNING: couldn't create faces for brush %i in entity %i (incomplete brush?)\n", brushnum, CurrentEntity - entities);
+		printf ("WARNING: couldn't create faces for brush %i in entity %i (incomplete brush?)\n", brushnum, (int)(CurrentEntity - entities));
 		return NULL;
 	}
